#5 refactor : chart그리기 로직 분리
